(chore) Add UI Playwright Tests to E2E Workflow (#551)
* add playwright tests to e2e workflow
* add UI zarf deployment to e2e workflow

+++ src/leapfrogai_ui/tests/global.teardown.ts
@@ -4,13 +4,16 @@ import { deleteAllAssistants, deleteAssistantAvatars } from './helpers/assistant
 import { deleteAllTestThreadsWithApi } from './helpers/threadHelpers';
 import { deleteAllTestAPIKeys } from './helpers/apiHelpers';
 
-teardown('teardown', async ({ openAIClient }) => {
-  console.log('cleaning up...');
-  deleteAllGeneratedFixtureFiles();
-  await deleteAllTestFilesWithApi(openAIClient);
-  await deleteAllAssistants(openAIClient);
-  await deleteAllTestThreadsWithApi(openAIClient);
-  await deleteAssistantAvatars();
-  await deleteAllTestAPIKeys();
-  console.log('clean up complete');
-});
+// teardown not necessary in CI testing envs
+if (process.env.TEST_ENV !== 'CI') {
+  teardown('teardown', async ({ openAIClient }) => {
+    console.log('cleaning up...');
+    deleteAllGeneratedFixtureFiles();
+    await deleteAllTestFilesWithApi(openAIClient);
+    await deleteAllAssistants(openAIClient);
+    await deleteAllTestThreadsWithApi(openAIClient);
+    await deleteAssistantAvatars();
+    await deleteAllTestAPIKeys();
+    console.log('clean up complete');
+  });
+}
